src/opticscontrol/pyproject.toml
[project]
name = "opticscontrol"
version = "0.1.0"
description = "Control software for Elliptec motorized stages"
authors = [{ name = "Kimoon Han", email = "khan@kaist.ac.kr" }]
requires-python = ">=3.11"
dependencies = [
    "construct>=2.10.68",
    "construct-typing>=0.6.2",
    "libusb>=1.0.27.post4",
    "numpy>=2.0.2",
    "pyftdi>=0.56.0",
    "pyqt6>=6.8.0",
    "pyusb>=1.3.1",
    "qtpy>=2.4.2",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
